Former-commit-id: c3685e6e1fb0d71078febafd44ae341f574d371e

+++ fortscale/presidio-output/presidio-output-webapp/src/main/java/presidio/webapp/dto/ListResponseBean.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class ListResponseBean<T> extends ResponseBean implements Serializable {
 
     private List<T> data;
-    private int total;
+    private long total;
     private int page;
 
     public ListResponseBean() {
@@ -27,11 +27,11 @@ public class ListResponseBean<T> extends ResponseBean implements Serializable {
         this.data = data;
     }
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
     }
